Refactor rendering code into functions
